requirements.txt
alembic==1.14.0
bidict==0.23.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.10.1
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
